+++ src/Field/Addresses.php
@@ -649,6 +649,7 @@ JS, [
             'label' => mb_ucfirst(t('Copy all {type}', [
                 'type' => Address::pluralLowerDisplayName(),
             ])),
+            'showInChips' => false,
         ];
     }